// src/protogen/event.pb.ts
export enum EventStatus {
  EVENT_STATUS_UNSPECIFIED = 0,
  EVENT_STATUS_DRAFT = 1,
  EVENT_STATUS_PUBLISHED = 2,
  EVENT_STATUS_CANCELLED = 3,
  EVENT_STATUS_CONFIGURED = 4,
  EVENT_STATUS_APPROVED = 5,
  UNRECOGNIZED = -1,
}

export enum EventRoleType {
  EVENT_ROLE_TYPE_UNSPECIFIED = 0,
  EVENT_ROLE_TYPE_ADMIN = 1,
  EVENT_ROLE_TYPE_EDITOR = 2,
  EVENT_ROLE_TYPE_VIEWER = 3,
  UNRECOGNIZED = -1,
}

export interface Event {
  id: string;
  name: string;
  description: string;
  startDate: string;
  endDate: string;
  thumbnailUrl: string;
  location: EventLocation | undefined;
  config: EventConfig | undefined;
  organizer: EventOrganizer | undefined;
  roles: EventRole[];
  categories: EventCategory[];
  status: EventStatus;
  createdAt: string;
  updatedAt: string;
}

export interface EventCategory {
  id: string;
  name: string;
}

export interface EventConfig {
  id: string;
  ticketSaleStartDate: string;
  ticketSaleEndDate: string;
  isFree: boolean;
  maxAttendees: number;
  isPublic: boolean;
  requiresApproval: boolean;
  allowWaitRoom: boolean;
  isNewTrending: boolean;
}

export interface EventLocation {
  id: string;
  venue: string;
  address: string;
}

export interface EventOrganizer {
  id: string;
  name: string;
  description: string;
  logoUrl: string;
}

export interface EventRole {
  id: string;
  userId: string;
  eventId: string;
  role: EventRoleType;
}